test(challenges): Corrected a test case for challenge
For the last test case in 'Access Props using this.props' challenge, I updated the test condition to check that the value of the tempPassword prop is the same as the content within the <strong/> tags. This test case will ensure that there is text present in the strong tag and its value is as expected.

BREAKING CHANGE: none

Closes #17134